refactor app composition helpers
